backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
